move `Scandir` to submodule
